// src/app/model/genre.ts
export enum Genre {
    DOCUMENTAL = 'Documental',
    SCI_FI = 'Sci-Fi',
    FANTASY = 'Fantasy',
    DRAMA = 'Drama',
    COMEDY = 'Comedy',
    BIOGRAPHY = 'Biography',
    CRIME = 'Crime',
    CARTOON = 'Cartoon',
    ANIME = 'Anime',
    ACTION = 'Action',
    ADVENTURE = 'Adventure',
    HORROR = 'Horror',
    HISTORY = 'History'
}